chore(emails): create email service
