+++ apps/backend/src/email/email.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { EmailSignUpService } from './email-sign-up/email-sign-up.service';
+import { EmailService } from './email.service';
 
 @Module({
-  providers: [EmailSignUpService],
-  exports: [EmailSignUpService],
+  providers: [EmailService],
+  exports: [EmailService],
 })
 export class EmailModule {}
